+++ editor/editor-core/src/__tests__/integration/paste/paste.ts
@@ -168,10 +168,14 @@ BrowserTestCase(
   },
 );
 
+// FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16630
 BrowserTestCase(
   'paste.ts: paste tests on fullpage editor with restart numbered lists: ordered list',
   // TODO: Re-skipped safari, need to unskip again after fixing: https://product-fabric.atlassian.net/browse/ED-16306
-  { skip: ['safari'] },
+  {
+    // skip: ['safari'],
+    skip: ['*'],
+  },
   async (client: any, testName: string) => {
     const page = await goToEditorTestingWDExample(client);
 
@@ -195,10 +199,14 @@ BrowserTestCase(
   },
 );
 
+// FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16631
 BrowserTestCase(
   'paste.ts: paste tests on fullpage editor: block node containing paragraph containing hardbreak and list',
   // TODO: Re-skipped safari, need to unskip again after fixing: https://product-fabric.atlassian.net/browse/ED-16306
-  { skip: ['safari'] },
+  {
+    // skip: ['safari'],
+    skip: ['*'],
+  },
   async (client: any, testName: string) => {
     const page = await goToEditorTestingWDExample(client);
 
@@ -220,10 +228,14 @@ BrowserTestCase(
   },
 );
 
+// FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16633
 BrowserTestCase(
   'paste.ts: code block copied from renderer and pasted',
   // TODO: Re-skipped safari, need to unskip again after fixing: https://product-fabric.atlassian.net/browse/ED-16306
-  { skip: ['safari'] },
+  {
+    // skip: ['safari'],
+    skip: ['*'],
+  },
   async (client: WebdriverIO.BrowserObject, testName: string) => {
     const page = await goToEditorTestingWDExample(client);
 
@@ -294,11 +306,15 @@ BrowserTestCase(
   },
 );
 
+// FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16634
 BrowserTestCase(
   'paste.ts: paste code with new lines into code block. ensure up arrow is working',
   // skip safari this test is for windows only
   // firefox skipped due to flaky copy paste https://product-fabric.atlassian.net/browse/ED-15079
-  { skip: ['safari', 'firefox'] },
+  {
+    // skip: ['safari', 'firefox'],
+    skip: ['*'],
+  },
   async (client: WebdriverIO.BrowserObject, testName: string) => {
     const page = await goToEditorTestingWDExample(client);