+++ editor/editor-core/src/__tests__/integration/toolbar/toolbar-3.ts
@@ -62,9 +62,12 @@ const clear = `span=${toolbarMessages.clearFormatting.defaultMessage}`;
 });
 
 [comment, fullpage].forEach((editor) => {
+  // FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16635
   BrowserTestCase(
     `toolbar-3.ts: should be able to select Clear Formatting on toolbar for ${editor.name} editor for node-based formatting`,
-    { skip: [] },
+    {
+      skip: ['*'],
+    },
     async (client: any, testName: string) => {
       const page = await goToEditorTestingWDExample(client);
       await mountEditor(page, {