+++ editor/editor-core/src/__tests__/visual-regression/common/panel.ts
@@ -294,7 +294,8 @@ describe('Dark mode panel', () => {
     await snapshot(page);
   });
 
-  it('Should render standarad panels', async () => {
+  // FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16629
+  it.skip('Should render standarad panels', async () => {
     adfContent = panel;
     await initFullPageEditorWithAdf(
       page,