+++ editor/editor-core/src/plugins/find-replace/__tests__/integration/match-case.ts
@@ -65,9 +65,12 @@ BrowserTestCase(
   },
 );
 
+// FIXME: This test was automatically skipped due to failure on 24/01/2023: https://product-fabric.atlassian.net/browse/ED-16637
 BrowserTestCase(
   'match-case.ts: find with Match Case button toggled on should find results that exactly match case',
-  {},
+  {
+    skip: ['*'],
+  },
   async (client: any, testName: string) => {
     const { page } = await openFindReplace(client, matchCaseAdf, true);
     await toggleMatchCase(page, true);